+++ docs/content/misc/release_notes.rst
@@ -10,6 +10,7 @@ Version 0.4.6
     - The list of Us4R class constructor parameters has been changed from list of positional \
       parameters to a list of pairs (name, value). E.g. use \
       ``Us4R('nArius', 2, 'voltage', 20, 'probeName', 'SL1543', 'adapterType', 'esaote3')``.
+    - Added ``dynamicRange`` to ``BModeDisplay`` class.
 
 Version 0.4.5
 -------------